Fix printing of mac addresses
On NUC, I was getting "ffffffb8:ffffffae:fffffffed...". Casting to
uint8_t before printing does the right thing.

+++ ports/fuchsia/netif/ethernetif.c
@@ -223,7 +223,8 @@ low_level_input(struct netif *netif)
       memcpy(netif->hwaddr, buf, 6);
       printf("ethernetif: link up (mac %02x:%02x:%02x:%02x:%02x:%02x,"
              " mtu %u)\n",
-             buf[0], buf[1], buf[2], buf[3], buf[4], buf[5], mtu);
+             (uint8_t)buf[0], (uint8_t)buf[1], (uint8_t)buf[2], (uint8_t)buf[3],
+             (uint8_t)buf[4], (uint8_t)buf[5], mtu);
       netif_set_link_up(netif);
     } else {
       /* link down */
